# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VSM Mobile",
        "description": "Documentation API VSM PK11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/announcements": {
            "get": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Liste de tous les communiqués",
                "description": "Récupère la liste des communiqués triés du plus récent au plus ancien.",
                "operationId": "018442d06de1537ba02109efdadd7a51",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "author_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Réunion générale"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Ordre du jour : Préparation du prochain match."
                                            },
                                            "isUrgent": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-08-28T14:30:00.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-08-28T14:30:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Créer un nouveau communiqué",
                "description": "Enregistre un nouveau communiqué en base de données.",
                "operationId": "aafebd32889be036f317a8ec262356ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Cotisation mensuelle",
                                        "maxLength": 255
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Prière de régler vos cotisations avant le 5 du mois."
                                    },
                                    "isUrgent": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Communiqué créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "author_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Cotisation mensuelle"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Prière de régler vos cotisations avant le 5 du mois."
                                        },
                                        "isUrgent": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation des données"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/announcements/{id}": {
            "put": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Mettre à jour un communiqué",
                "description": "Modifie les informations d'un communiqué existant.",
                "operationId": "fc9e73377f5093b3a925537950ccafd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du communiqué",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mise à jour : Réunion générale",
                                        "maxLength": 255
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "La réunion est reportée à 16h00."
                                    },
                                    "isUrgent": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Communiqué mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "author_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Mise à jour : Réunion générale"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "La réunion est reportée à 16h00."
                                        },
                                        "isUrgent": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Communiqué non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation des données"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Supprimer un communiqué",
                "description": "Supprime définitivement un communiqué de la base de données.",
                "operationId": "83a5ea4614797a52fe1f4390c5743bd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du communiqué",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Communiqué supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Communiqué supprimé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Communiqué non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrez votre jeton Sanctum sous la forme: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}